[IMP] hr_course: Add improvements to 14
